Refactor check_for_loop_range into its module
